Move TextPropertyWidget to separate directory.
